+++ src/main/java/de/maxhenkel/corpse/entities/CorpseRenderer.java
@@ -1,71 +1,65 @@
 package de.maxhenkel.corpse.entities;
 
+import com.mojang.authlib.GameProfile;
 import com.mojang.blaze3d.matrix.MatrixStack;
+import de.maxhenkel.corelib.CachedMap;
 import de.maxhenkel.corelib.client.PlayerSkins;
 import de.maxhenkel.corpse.Main;
+import net.minecraft.client.entity.player.AbstractClientPlayerEntity;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
 import net.minecraft.client.renderer.entity.EntityRenderer;
 import net.minecraft.client.renderer.entity.EntityRendererManager;
-import net.minecraft.client.renderer.entity.model.PlayerModel;
-import net.minecraft.client.renderer.entity.model.SkeletonModel;
-import net.minecraft.entity.Entity;
+import net.minecraft.client.renderer.entity.SkeletonRenderer;
+import net.minecraft.client.world.ClientWorld;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.vector.Vector3f;
 
 public class CorpseRenderer extends EntityRenderer<CorpseEntity> {
 
-    private static final ResourceLocation SKELETON_TEXTURE = new ResourceLocation("textures/entity/skeleton/skeleton.png");
-
-    private PlayerModel modelPlayer;
-    private PlayerModel modelPlayerSlim;
-    private SkeletonModel modelSkeleton;
+    private final CachedMap<CorpseEntity, DummyPlayer> players;
+    private final CachedMap<CorpseEntity, DummySkeleton> skeletons;
+    private final DummyPlayerRenderer playerRenderer;
+    private final DummyPlayerRenderer playerRendererSmallArms;
+    private final SkeletonRenderer skeletonRenderer;
 
     public CorpseRenderer(EntityRendererManager renderManager) {
         super(renderManager);
-        modelPlayer = new PlayerModel(0F, false);
-        modelPlayerSlim = new PlayerModel(0F, true);
-        modelSkeleton = new SkeletonModel() {
-            @Override
-            public void setRotationAngles(Entity entityIn, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
-
-            }
-        };
-        modelPlayer.isChild = false;
-        modelPlayerSlim.isChild = false;
-        modelSkeleton.isChild = false;
+        players = new CachedMap<>(10_000L);
+        skeletons = new CachedMap<>(10_000L);
+        playerRenderer = new DummyPlayerRenderer(renderManager, false);
+        playerRendererSmallArms = new DummyPlayerRenderer(renderManager, true);
+        skeletonRenderer = new SkeletonRenderer(renderManager);
     }
 
     @Override
     public ResourceLocation getEntityTexture(CorpseEntity entity) {
-        if (entity.isSkeleton()) {
-            return SKELETON_TEXTURE;
-        } else {
-            return PlayerSkins.getSkin(entity.getCorpseUUID(), entity.getCorpseName());
-        }
+        return null;
     }
 
     @Override
-    public void render(CorpseEntity entity, float f1, float f2, MatrixStack matrixStack, IRenderTypeBuffer buffer, int i) {
-        super.render(entity, f1, f2, matrixStack, buffer, i);
+    public void render(CorpseEntity entity, float entityYaw, float partialTicks, MatrixStack matrixStack, IRenderTypeBuffer buffer, int packedLightIn) {
+        super.render(entity, entityYaw, partialTicks, matrixStack, buffer, packedLightIn);
         matrixStack.push();
 
-        matrixStack.rotate(Vector3f.YP.rotationDegrees(360F - entity.getCorpseRotation()));
+        matrixStack.rotate(Vector3f.YP.rotationDegrees(-entity.getCorpseRotation()));
 
         if (Main.SERVER_CONFIG.spawnCorpseOnFace.get()) {
-            matrixStack.rotate(Vector3f.XP.rotationDegrees(-90F));
-            matrixStack.translate(0D, -0.5D, 2.01D / 16D);
-        } else {
             matrixStack.rotate(Vector3f.XP.rotationDegrees(90F));
-            matrixStack.translate(0D, -0.5D, -2.01D / 16D);
+            matrixStack.translate(0D, -1D, -2.01D / 16D);
+        } else {
+            matrixStack.rotate(Vector3f.XP.rotationDegrees(-90F));
+            matrixStack.translate(0D, -1D, 2.01D / 16D);
         }
 
         if (entity.isSkeleton()) {
-            modelSkeleton.render(matrixStack, buffer.getBuffer(modelSkeleton.getRenderType(getEntityTexture(entity))), i, 0xFFFFFF, 1F, 1F, 1F, 1F);
+            DummySkeleton skeleton = skeletons.get(entity, () -> new DummySkeleton(entity.world, entity.getEquipment()));
+            skeletonRenderer.render(skeleton, entityYaw, 1F, matrixStack, buffer, packedLightIn);
         } else {
+            AbstractClientPlayerEntity abstractClientPlayerEntity = players.get(entity, () -> new DummyPlayer((ClientWorld) entity.world, new GameProfile(entity.getCorpseUUID(), entity.getCorpseName()), entity.getEquipment()));
             if (PlayerSkins.isSlim(entity.getCorpseUUID())) {
-                modelPlayerSlim.render(matrixStack, buffer.getBuffer(modelPlayerSlim.getRenderType(getEntityTexture(entity))), i, 0xFFFFFF, 1F, 1F, 1F, 1F);
+                playerRendererSmallArms.render(abstractClientPlayerEntity, 0F, 1F, matrixStack, buffer, packedLightIn);
             } else {
-                modelPlayer.render(matrixStack, buffer.getBuffer(modelPlayer.getRenderType(getEntityTexture(entity))), i, 0xFFFFFF, 1F, 1F, 1F, 1F);
+                playerRenderer.render(abstractClientPlayerEntity, 0F, 1F, matrixStack, buffer, packedLightIn);
             }
         }
         matrixStack.pop();